Organize code into packages, fix design doc
Split flat main package into six packages under pkg/ with compile-time
boundary enforcement:

  pkg/worktree  — Entry model, sorting, name generation
  pkg/ssh       — remote execution, home resolution, path translation
  pkg/git       — repo root, worktree creation, dir existence
  pkg/opencode  — session discovery (HTTP API + SQLite), enrichment
  pkg/discover  — local filesystem walk, remote SSH discovery
  pkg/display   — table formatting

Library code returns errors instead of calling die(). Process exec
functions stay in main (only consumer, no type at interface).

Design doc updates:
- Fix wt ls: document actual 6-column merged table, column semantics,
  status states (working/idle), dash meaning, sort order
- Fix Model section: working/idle instead of stale idle/busy
- Fix stale opencode --continue reference (flag doesn't exist)
- Fix Implementation section: mention both SQLite and HTTP API
- Repo display: shorten to <home>/.../parent/name, [remote] tag
- Scrub real repo/user names from examples
- Move design to docs/designs/wt.md
