+++ docs/designs/wt.md
@@ -24,10 +24,13 @@ equals the branch name. The worktree directory is the primary key. Worktrees are
 created per unit of work and cleaned up separately when the branch lands.
 
 A **session** is an OpenCode conversation bound to a worktree directory. Sessions
-persist on the server, carry a title (auto-generated from the first prompt), a
-status (idle/busy), and full message history. A worktree can have multiple
-sessions; the most recent is attached by default. Session lifecycle (creating new
-sessions, forking) is managed from within OpenCode, not by this tool.
+persist on the server and carry a title (auto-generated from the first prompt)
+and full message history. A worktree can have multiple sessions; the most recent
+is attached by default. Session lifecycle (creating new sessions, forking) is
+managed from within OpenCode, not by this tool.
+
+A session is **working** when the agent is actively generating a response, and
+**idle** otherwise. A worktree with no session has no status.
 
 The **OpenCode server** is a persistent process on the dev desktop. One server
 hosts all worktrees across all repos. Every API endpoint accepts a `directory`
@@ -54,7 +57,7 @@ through an SSH tunnel.
 
 ```
 laptop                                  dev desktop
-  wt -r ~/go/src/.../kro ──SSH──>         git worktree add ...
+  wt -r ~/src/acme/api ──SSH──>            git worktree add ...
                                                 │
   opencode attach ────tunnel────────> opencode serve
     --dir <remote worktree path>
@@ -72,14 +75,14 @@ Create or resume a worktree.
 - No args: create a new worktree in the current repo. Attach.
 - With `name`: resume `<repo>/.worktrees/<name>`. Attach.
 
-Attach: `opencode --continue` in the worktree directory.
+Attach: run `opencode` in the worktree directory.
 
 ### `wt -r <path> [name]`
 
 Create or resume a remote worktree.
 
 - `path` identifies the repo as a local-style path
-  (e.g., `~/go/src/github.com/ellistarn/kro`), translated to the remote
+  (e.g., `~/src/acme/api`), translated to the remote
   equivalent.
 - Without `name`: create a new worktree. Attach.
 - With `name`: resume existing worktree. Attach.
@@ -90,27 +93,37 @@ session ID.
 
 ### `wt ls`
 
-List all worktrees and their session status. Shows both local (current repo) and
-remote (all repos on the dev desktop). If the remote is unreachable, shows local
-only.
+List all worktrees and their session status. Local worktrees (all repos under
+`$HOME`) and remote worktrees (all repos on the dev desktop) are discovered
+concurrently and merged into a single table sorted by most recent activity.
 
 ```
-LOCAL
-  WORKTREE          STATUS   TITLE
-  0423T1600-4419    idle     Refactor config parser
-
-REMOTE
-  WORKTREE          REPO          STATUS   TITLE
-  0423T1430-12847   .../kro       busy     Fix auth handler validation
-  0422T0900-9182    .../karp      idle     Add node pool autoscaling
-  0421T1100-5531    .../ark       -        -
+  UPDATED     CREATED     WORKTREE            REPO                                STATUS    TITLE
+  just now    3h ago      0423T1430-12847     [remote] /home/user/.../acme/api    working   Fix auth handler validation
+  5m ago      1d ago      0423T1600-4419      /Users/user/.../acme/api            idle      Refactor config parser
+  -           2d ago      0421T1100-5531      [remote] /home/user/.../acme/web    -         -
 ```
+
+Columns:
+
+| Column | Value |
+|--------|-------|
+| UPDATED | When the most recent session in this worktree was last active. |
+| CREATED | When the worktree was created. |
+| WORKTREE | Branch name (equals the worktree directory name). |
+| REPO | Repo root, shortened to `<home>/.../parent/name`. `[remote]` prefix for remote worktrees. |
+| STATUS | `working` (agent generating), `idle` (session exists, agent not generating). |
+| TITLE | Session title, auto-generated from the first prompt. |
+
+`-` in any column means the value is unavailable. A worktree with no session
+shows `-` for UPDATED, STATUS, and TITLE. Attaching to such a worktree creates
+a session; subsequent listings show its status and title.
 
 ## Reconnection
 
 1. Laptop opens. SSH tunnel restarts.
 2. `wt ls` shows everything in flight.
-3. `wt -r ~/go/src/.../kro 0423T1430-12847` resumes.
+3. `wt -r ~/src/acme/api 0423T1430-12847` resumes.
 
 ## Assumptions
 
@@ -121,8 +134,9 @@ REMOTE
 ## Implementation
 
 Go binary. Shells out to `ssh` for remote git operations and to `opencode` for
-TUI attachment. Talks to the OpenCode server via `net/http` for session discovery.
-Single `main` package.
+TUI attachment. Queries the OpenCode SQLite database for session metadata in
+listings. Queries the OpenCode HTTP API for session discovery when attaching
+remotely.
 
 ## Scoped Out
