+++ pkg/display/display.go
@@ -1,13 +1,17 @@
-package main
+package display
 
 import (
 	"fmt"
 	"os"
+	"strings"
 	"text/tabwriter"
 	"time"
+
+	"github.com/ellistarn/wt/pkg/worktree"
 )
 
-func printWorktreeTable(entries []WorktreeEntry) {
+// PrintTable prints worktree entries as an aligned table.
+func PrintTable(entries []worktree.Entry) {
 	w := tabwriter.NewWriter(os.Stdout, 2, 4, 2, ' ', 0)
 
 	fmt.Fprintf(w, "  UPDATED\tCREATED\tWORKTREE\tREPO\tSTATUS\tTITLE\n")
@@ -31,10 +35,18 @@ func printWorktreeTable(entries []WorktreeEntry) {
 	w.Flush()
 }
 
-// formatRepo prefixes remote repos with ssh:// to distinguish them from local.
+// formatRepo shortens the repo path to <home>/.../last/two and tags remote entries.
 func formatRepo(repo string, remote bool) string {
+	parts := strings.Split(repo, "/")
+	// Shorten: keep first 3 segments (e.g., "", "Users", "etarn"), then .../, then last 2.
+	// Only shorten if there are more than 5 segments (home + at least 3 intermediate + 2 tail).
+	if len(parts) > 5 {
+		head := strings.Join(parts[:3], "/") // e.g., /Users/etarn
+		tail := strings.Join(parts[len(parts)-2:], "/")
+		repo = head + "/.../" + tail
+	}
 	if remote {
-		return "ssh://" + repo
+		return "[remote] " + repo
 	}
 	return repo
 }
