+++ pkg/worktree/worktree.go
@@ -1,4 +1,4 @@
-package main
+package worktree
 
 import (
 	"fmt"
@@ -7,31 +7,32 @@ import (
 	"time"
 )
 
-// WorktreeEntry represents a discovered worktree.
-type WorktreeEntry struct {
+// Entry represents a discovered worktree.
+type Entry struct {
 	Name      string    // branch/worktree name
 	Dir       string    // absolute path on the host where it lives
 	Repo      string    // repo root path
 	Remote    bool      // true if this worktree lives on the remote host
 	CreatedAt time.Time // worktree creation time (from filesystem)
-	UpdatedAt time.Time // last session activity (from OpenCode server)
+	UpdatedAt time.Time // last session activity (from OpenCode)
 	SessionID string    // most recent OpenCode session ID (empty if none)
-	Status    string    // working, idle, or missing
+	Status    string    // working or idle; empty if no session
 	Title     string    // OpenCode session title
 }
 
-func generateName() string {
+// GenerateName returns a timestamped random name for a new worktree.
+func GenerateName() string {
 	now := time.Now()
 	return fmt.Sprintf("%s-%d", now.Format("0102T1504"), rand.Intn(100000))
 }
 
-// sortWorktrees sorts entries by most recent activity, newest first.
+// Sort sorts entries by most recent activity, newest first.
 // Uses UpdatedAt if available, otherwise CreatedAt.
 // Entries without any timestamp sort to the end.
-func sortWorktrees(entries []WorktreeEntry) {
+func Sort(entries []Entry) {
 	sort.SliceStable(entries, func(i, j int) bool {
-		ti := entries[i].sortTime()
-		tj := entries[j].sortTime()
+		ti := sortTime(entries[i])
+		tj := sortTime(entries[j])
 		if ti.IsZero() && tj.IsZero() {
 			return entries[i].Name < entries[j].Name
 		}
@@ -45,15 +46,14 @@ func sortWorktrees(entries []WorktreeEntry) {
 	})
 }
 
-// sortTime returns UpdatedAt if set, otherwise CreatedAt.
-func (e *WorktreeEntry) sortTime() time.Time {
+func sortTime(e Entry) time.Time {
 	if !e.UpdatedAt.IsZero() {
 		return e.UpdatedAt
 	}
 	return e.CreatedAt
 }
 
-// timeUnix converts a unix timestamp to time.Time.
-func timeUnix(sec int64) time.Time {
+// TimeUnix converts a unix timestamp to time.Time.
+func TimeUnix(sec int64) time.Time {
 	return time.Unix(sec, 0)
 }
